+++ modules/apps/app-builder/app-builder-web/src/main/resources/META-INF/resources/js/pages/apps/ListApps.es.js
@@ -13,13 +13,13 @@
  */
 
 import ClayLabel from '@clayui/label';
+import ListView from 'data-engine-js-components-web/js/components/list-view/ListView.es';
 import {compile} from 'path-to-regexp';
 import React, {useContext} from 'react';
 import {Link} from 'react-router-dom';
 
 import {AppContext} from '../../AppContext.es';
 import Button from '../../components/button/Button.es';
-import ListView from '../../components/list-view/ListView.es';
 import useBackUrl from '../../hooks/useBackUrl.es';
 import useDataDefinition from '../../hooks/useDataDefinition.es';
 import useDeployApp from '../../hooks/useDeployApp.es';
@@ -60,7 +60,7 @@ export const Actions = () => {
 	];
 };
 
-export default ({
+const ListApps = ({
 	editPath = [
 		`/:objectType/:dataDefinitionId(\\d+)/apps/deploy`,
 		`/:objectType/:dataDefinitionId(\\d+)/apps/:appId(\\d+)`,
@@ -69,6 +69,7 @@ export default ({
 	match: {
 		params: {dataDefinitionId, objectType},
 	},
+	history,
 }) => {
 	const {scope} = useContext(AppContext);
 	const withBackUrl = useBackUrl();
@@ -117,6 +118,7 @@ export default ({
 			columns={COLUMNS}
 			emptyState={EMPTY_STATE}
 			endpoint={ENDPOINT}
+			history={history}
 			{...listViewProps}
 		>
 			{(app) => {
@@ -145,3 +147,5 @@ export default ({
 		</ListView>
 	);
 };
+
+export default ListApps;
