+++ modules/apps/app-builder/app-builder-web/src/main/resources/META-INF/resources/js/pages/entry/ListEntries.es.js
@@ -12,11 +12,11 @@
  * details.
  */
 
+import ListView from 'data-engine-js-components-web/js/components/list-view/ListView.es';
 import React, {useContext} from 'react';
 
 import {AppContext} from '../../AppContext.es';
 import Button from '../../components/button/Button.es';
-import ListView from '../../components/list-view/ListView.es';
 import {Loading} from '../../components/loading/Loading.es';
 import useDataListView from '../../hooks/useDataListView.es';
 import useEntriesActions from '../../hooks/useEntriesActions.es';
@@ -107,6 +107,7 @@ export default function ListEntries({history}) {
 					title: Liferay.Language.get('there-are-no-entries-yet'),
 				}}
 				endpoint={`/o/data-engine/v2.0/data-definitions/${dataDefinitionId}/data-records`}
+				history={history}
 				noActionsMessage={Liferay.Language.get(
 					'you-do-not-have-the-permission-to-manage-this-entry'
 				)}
