+++ modules/apps/app-builder/app-builder-web/src/main/resources/META-INF/resources/js/pages/object/ListObjects.es.js
@@ -12,12 +12,12 @@
  * details.
  */
 
+import ListView from 'data-engine-js-components-web/js/components/list-view/ListView.es';
 import React, {useEffect, useState} from 'react';
 import {Link} from 'react-router-dom';
 
 import {AppNavigationBar} from '../../App.es';
 import ControlMenu from '../../components/control-menu/ControlMenu.es';
-import ListView from '../../components/list-view/ListView.es';
 import PermissionsModal from '../../components/permissions/PermissionsModal.es';
 import {ACTIONS} from '../../pages/entry/PermissionsContext.es';
 import {getItem, updateItem} from '../../utils/client.es';
@@ -125,7 +125,12 @@ export default ({history, listViewProps = {}, objectType}) => {
 
 			<AppNavigationBar />
 
-			<ListView {...listViewProps} actions={actions} columns={columns}>
+			<ListView
+				{...listViewProps}
+				actions={actions}
+				columns={columns}
+				history={history}
+			>
 				{(item) => {
 					const {
 						dateCreated,
