+++ modules/apps/app-builder/app-builder-web/src/main/resources/META-INF/resources/js/pages/table-view/ListTableViews.es.js
@@ -12,11 +12,11 @@
  * details.
  */
 
+import ListView from 'data-engine-js-components-web/js/components/list-view/ListView.es';
 import React from 'react';
 import {Link} from 'react-router-dom';
 
 import Button from '../../components/button/Button.es';
-import ListView from '../../components/list-view/ListView.es';
 import useDataDefinition from '../../hooks/useDataDefinition.es';
 import {confirmDelete} from '../../utils/client.es';
 import {getLocalizedValue} from '../../utils/lang.es';
@@ -86,6 +86,7 @@ export default ({
 				title: Liferay.Language.get('there-are-no-table-views-yet'),
 			}}
 			endpoint={`/o/data-engine/v2.0/data-definitions/${dataDefinitionId}/data-list-views`}
+			history={history}
 		>
 			{(item) => {
 				const {dateCreated, dateModified, id, name} = item;
